Refactor after rebasing with master
