+++ pinot-query-planner/src/main/java/org/apache/pinot/query/planner/logical/RexExpression.java
@@ -19,11 +19,8 @@
 package org.apache.pinot.query.planner.logical;
 
 import com.google.common.base.Preconditions;
-import com.google.common.collect.Range;
 import java.math.BigDecimal;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Set;
 import java.util.stream.Collectors;
 import org.apache.calcite.rel.core.AggregateCall;
 import org.apache.calcite.rel.type.RelDataType;
@@ -33,8 +30,6 @@ import org.apache.calcite.rex.RexLiteral;
 import org.apache.calcite.rex.RexNode;
 import org.apache.calcite.sql.SqlKind;
 import org.apache.calcite.util.NlsString;
-import org.apache.calcite.util.Sarg;
-import org.apache.commons.lang3.NotImplementedException;
 import org.apache.pinot.common.utils.PinotDataType;
 import org.apache.pinot.query.planner.serde.ProtoProperties;
 import org.apache.pinot.spi.data.FieldSpec;
@@ -59,29 +54,29 @@ public interface RexExpression {
       return new RexExpression.Literal(dataType, toRexValue(dataType, rexLiteral.getValue()));
     } else if (rexNode instanceof RexCall) {
       RexCall rexCall = (RexCall) rexNode;
-      if (rexCall.getKind().equals(SqlKind.SEARCH)) {
-        return handleSearch(rexCall);
-      }
-      List<RexExpression> operands = rexCall.getOperands().stream().map(RexExpression::toRexExpression)
-          .collect(Collectors.toList());
-      return toRexExpression(rexCall, operands);
+      return toRexExpression(rexCall);
     } else {
       throw new IllegalArgumentException("Unsupported RexNode type with SqlKind: " + rexNode.getKind());
     }
   }
 
-  static RexExpression toRexExpression(RexCall rexCall, List<RexExpression> operands) {
+  static RexExpression toRexExpression(RexCall rexCall) {
+    List<RexExpression> operands;
     switch (rexCall.getKind()) {
       case CAST:
         // CAST is being rewritten into "rexCall.CAST<targetType>(inputValue)",
         //   - e.g. result type has already been converted into the CAST RexCall, so we assert single operand.
+        operands = rexCall.getOperands().stream().map(RexExpression::toRexExpression).collect(Collectors.toList());
         Preconditions.checkState(operands.size() == 1, "CAST takes exactly 2 arguments");
         RelDataType castType = rexCall.getType();
         // add the 2nd argument as the source type info.
-        operands.add(new Literal(FieldSpec.DataType.STRING, rexCall.getOperands().get(0).getType().getSqlTypeName(),
+        operands.add(new Literal(FieldSpec.DataType.STRING,
             toPinotDataType(rexCall.getOperands().get(0).getType()).name()));
         return new RexExpression.FunctionCall(rexCall.getKind(), toDataType(rexCall.getType()), "CAST", operands);
+      case SEARCH:
+        return RexExpressionUtils.handleSearch(rexCall);
       default:
+        operands = rexCall.getOperands().stream().map(RexExpression::toRexExpression).collect(Collectors.toList());
         return new RexExpression.FunctionCall(rexCall.getKind(), toDataType(rexCall.getType()),
             rexCall.getOperator().getName(), operands);
     }
@@ -150,33 +145,6 @@ public interface RexExpression {
         // TODO: do not assume byte type.
         return FieldSpec.DataType.BYTES;
     }
-  }
-
-  // POC Code. Will clean-up later.
-  static RexExpression handleSearch(RexCall rexCall) {
-    List<RexNode> operands = rexCall.getOperands();
-    RexInputRef rexInputRef = (RexInputRef) operands.get(0);
-    RexLiteral rexLiteral = (RexLiteral) operands.get(1);
-    FieldSpec.DataType dataType = toDataType(rexLiteral.getType());
-    Sarg sarg = rexLiteral.getValueAs(Sarg.class);
-    if (sarg.isPoints()) {
-      return new FunctionCall(SqlKind.IN, dataType, SqlKind.IN.name(), convertRanges(rexInputRef,
-          sarg.rangeSet.asRanges(), dataType));
-    } else if (sarg.isComplementedPoints()) {
-      return new FunctionCall(SqlKind.NOT_IN, dataType, SqlKind.NOT_IN.name(), convertRanges(rexInputRef,
-          sarg.rangeSet.complement().asRanges(), dataType));
-    } else {
-      throw new NotImplementedException("Range is not implemented yet");
-    }
-  }
-
-  static List<RexExpression> convertRanges(RexInputRef rexInputRef, Set<Range> ranges, FieldSpec.DataType dataType) {
-    List<RexExpression> result = new ArrayList<>(ranges.size() + 1);
-    result.add(toRexExpression(rexInputRef));
-    for (Range range : ranges) {
-      result.add(new Literal(dataType, toRexValue(dataType, range.lowerEndpoint())));
-    }
-    return result;
   }
 
   class InputRef implements RexExpression {
